requirements.txt
alembic==1.16.2
altgraph==0.17.4
asgiref==3.8.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
dj-database-url==2.3.0
Django==5.2
django-allauth==65.9.0
django-channels==0.7.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-tailwind==4.0.1
djangorestframework==3.16.0
dnspython==2.7.0
email-validator==2.1.0
Flask==2.3.3
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
fonttools==4.55.0
greenlet==3.2.3
gunicorn==21.2.0
h11==0.16.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
oauthlib==3.3.1
packaging==24.2
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.13.0
pywin32-ctypes==0.2.3
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stripe==9.5.0
tk==0.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==2.3.7
whitenoise==6.6.0
wsproto==1.2.0
WTForms==3.2.1
